improvement: make link shortcode easier & fix shader anchors
We can now write `{% link fileName[.md] [text [text]] %}` which gets
converted into [text](path-to-fileName.html) if `text` is supplied.

Note that `text` can be a multi-word text, for example:
`{% link learn-to-code-how-to-ask-questions How to ask questions %}`
get replaced by:
`[How to ask questions](path-to-learn-to-code-how-to-ask-questions.html)`

Also fixed the include shortcode anchor regex so it properly detects
contents for `.shader` files.

See more info in the `gdquest-build-course -h` help message.